+++ workers/ohsome_quality_analyst/base/indicator.py
@@ -63,7 +63,6 @@ class BaseIndicator(metaclass=ABCMeta):
         self,
         layer_name: str,
         feature: Feature,
-        time_range=str,
     ) -> None:
         self.feature = feature