+++ workers/ohsome_quality_analyst/cli/cli.py
@@ -104,7 +104,6 @@ def get_indicator_layer_combination():
 @cli_option(options.dataset_name)
 @cli_option(options.feature_id)
 @cli_option(options.fid_field)
-@cli_option(options.time_range)
 @cli_option(options.force)
 def create_indicator(
     indicator_name: str,
@@ -114,7 +113,6 @@ def create_indicator(
     feature_id: str,
     dataset_name: str,
     fid_field: str,
-    time_range: str,
     force: bool,
 ):
     """Create an Indicator.
@@ -141,7 +139,6 @@ def create_indicator(
             dataset_name,
             feature_id,
             fid_field,
-            time_range,
             force,
         )
     )
@@ -170,7 +167,6 @@ def create_report(
     dataset_name: str,
     feature_id: str,
     fid_field: str,
-    time_range: str,
     force: bool,
 ):
     """Create a Report.
@@ -196,7 +192,6 @@ def create_report(
             dataset_name,
             feature_id,
             fid_field,
-            time_range,
             force,
         )
     )